requirements.txt
asgiref
boto3
dj-database-url
django~=3.0.7
django-heroku
gunicorn
Jinja2
MarkupSafe
Pillow
psycopg2
pytz
six
sqlparse
django-storages
whitenoise
djangorestframework~=3.11.0
setuptools~=41.1.0